pkg/asset/cluster: move tfvars to its own package
Moves tfvars to its own package to decouple it from the cluster asset.
