+++ pkg/asset/cluster/tfvars/tfvars.go
@@ -1,4 +1,4 @@
-package cluster
+package tfvars
 
 import (
 	"context"
